ajoute les compétences fortes au retour de l'api
